requirements.txt
discord.py
python-dotenv
translate